+++ index.html
@@ -234,8 +234,8 @@
       <h2 class="data-error__title">Не удалось загрузить данные</h2>
     </section>
   </template>
-  <script src="functions.js"></script>
-  <script src="./js/main.js"></script>
+  <!-- <script src="functions.js"></script> -->
+  <script src="./js/main.js" type="module"></script>
 </body>
 
 </html>
