+++ js/data.js
@@ -1,0 +1,31 @@
+//Массив с возможными сообщениями комментариев
+const MESSAGES = [
+  'Всё отлично!',
+  'В целом всё неплохо. Но не всё.',
+  'Когда вы делаете фотографию, хорошо бы убирать палец из кадра. В конце концов это просто непрофессионально.',
+  'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.',
+  'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.',
+  'Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент?!'
+];
+
+//Массив с возможными именами комментаторов
+const NAMES = [
+  'Маша',
+  'Глаша',
+  'Вася',
+  'Петруша',
+  'Котя',
+  'Труся',
+  'Витя',
+  'Сержио'
+];
+
+const DESCRIPTIONS = [
+  'На отдыхе!',
+  'Мой кот обожает делать это!',
+  'Париж прекрасен в любую погоду.',
+  'Кто со мной на гонку героев?',
+  'Злободневно'
+];
+
+export{MESSAGES, NAMES, DESCRIPTIONS};
